Refactor RunApp and FileTester
In order to eliminate duplicate output, have all testers which work with files
use FileTester instead of RunApp. Move several file testing methods out of RunApp
which was being used by only a couple testers, into FileTester.

Adding a unittest for ImageDiff tester. Even though unittesting occurs in Chigger,
I felt having a simple test in the TestHarness wise.

Actually run arbitrary commands.

Closes #18974
