+++ python/TestHarness/testers/RunException.py
@@ -7,14 +7,15 @@
 #* Licensed under LGPL 2.1, please see LICENSE for details
 #* https://www.gnu.org/licenses/lgpl-2.1.html
 
+from FileTester import FileTester
+from RunApp import RunApp # for checkRunnable
 from TestHarness import util
-from RunApp import RunApp
 
-class RunException(RunApp):
+class RunException(FileTester):
 
     @staticmethod
     def validParams():
-        params = RunApp.validParams()
+        params = FileTester.validParams()
 
         params.addParam('expect_err', "A regular expression or literal string that must occur in the output (see match_literal). (Test may terminate unexpectedly and be considered passing)")
         params.addParam('expect_assert', "DEBUG MODE ONLY: A regular expression that must occur in the output. (Test may terminate unexpectedly and be considered passing)")
@@ -26,7 +27,7 @@ class RunException(RunApp):
         return params
 
     def __init__(self, name, params):
-        RunApp.__init__(self, name, params)
+        FileTester.__init__(self, name, params)
         if (params.isValid("expect_err") == False and params.isValid("expect_assert") == False):
             raise RuntimeError('Either "expect_err" or "expect_assert" must be supplied in RunException')
