+++ test/tests/test_harness/imagediffs
@@ -1,0 +1,16 @@
+[Tests]
+  [imagediff_ok]
+    type = ImageDiff
+    command = 'true'
+    imagediff = 'kitten.png'
+    display_required = false
+    delete_output_before_running = false
+  []
+  [imagediff_fail]
+    type = ImageDiff
+    command = 'true'
+    imagediff = 'not_a_kitten.png'
+    display_required = false
+    delete_output_before_running = false
+  []
+[]
