+++ test/tests/test_harness/kitten.png
@@ -1,0 +1,1 @@
+gold/kitten.png
