+++ test/tests/test_harness/not_a_kitten.png
@@ -1,0 +1,1 @@
+gold/kitten.png
